requirements.txt
allure-pytest         2.13.5
allure-python-commons 2.13.5
attrs                 23.2.0
certifi               2024.6.2
charset-normalizer    3.3.2
colorama              0.4.6
idna                  3.7
iniconfig             2.0.0
packaging             24.1
pip                   23.2.1
pluggy                1.5.0
pytest                8.2.2
requests              2.32.3
urllib3               2.2.2